requirements.txt
adal==1.2.7
applicationinsights==0.11.10
azure-common==1.1.27
azure-core==1.19.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.2.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-cli-common==1.36.0
azureml-core==1.33.0
azureml-pipeline-core==1.36.0
azureml-telemetry==1.36.0
azureml-train-core==1.36.0
azureml-train-restclients-hyperdrive==1.36.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
contextlib2==0.6.0.post1
cryptography==3.4.8
docker==4.4.4
idna==3.3
isodate==0.6.0
jeepney==0.7.1
jmespath==0.10.0
jsonpickle==2.0.0
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
oauthlib==3.1.1
pathspec==0.9.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
six==1.16.0
urllib3==1.26.5
websocket-client==1.2.1
